requirements.txt
click==8.0.3
dataclasses==0.8
flake8==4.0.1
Flask==2.0.2
gevent==1.2.0
greenlet==1.1.2
importlib-metadata==4.2.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.2
typing-extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
boto3==1.23.10
